+++ src/imdb-obsidian-base-watchlist/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-09-24
+
+### Added
+
+- Name, Created on added
+- description added
+
+### Fixed
+
+- Language, Country is properly calulated
+- Runtime is properly calulated in case of missing certification ratings
+- Category is properly calulated
+
 ## [1.0.0] - 2025-09-24
 
 ### Added